src/requirements.txt
numpy==1.26.1
pandas==2.1.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-env==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
